# Optimize SM2 on aarch64