--- a/crypto/ec/build.info
+++ b/crypto/ec/build.info
@@ -31,6 +31,13 @@ IF[{- !$disabled{asm} -}]
   $ECDEF_armv4=ECP_NISTZ256_ASM
   $ECASM_aarch64=ecp_nistz256.c ecp_nistz256-armv8.S
   $ECDEF_aarch64=ECP_NISTZ256_ASM
+  IF[{- !$disabled{'sm2'} -}]
+    $ECASM_aarch64=$ECASM_aarch64 ecp_sm2p256.c ecp_sm2p256-armv8.S
+    IF[{- !$disabled{'sm2-precomp'} -}]
+      $ECASM_aarch64=$ECASM_aarch64 ecp_sm2p256_table.c
+    ENDIF
+    $ECDEF_aarch64=$ECDEF_aarch64 ECP_SM2P256_ASM
+  ENDIF
 
   $ECASM_parisc11=
   $ECASM_parisc20_64=
@@ -124,4 +131,9 @@ GENERATE[ecp_nistp521-ppc64.s]=asm/ecp_nistp521-ppc64.pl
 IF[{- !$disabled{'ecx'} -}]
 GENERATE[x25519-x86_64.s]=asm/x25519-x86_64.pl
 GENERATE[x25519-ppc64.s]=asm/x25519-ppc64.pl
+ENDIF
+
+IF[{- !$disabled{'sm2'} -}]
+  GENERATE[ecp_sm2p256-armv8.S]=asm/ecp_sm2p256-armv8.pl
+  INCLUDE[ecp_sm2p256-armv8.o]=..
 ENDIF