--- a/crypto/ec/ec_local.h
+++ b/crypto/ec/ec_local.h
@@ -628,6 +628,11 @@ int ossl_ec_key_simple_generate_key(EC_KEY *eckey);
 int ossl_ec_key_simple_generate_public_key(EC_KEY *eckey);
 int ossl_ec_key_simple_check_key(const EC_KEY *eckey);
 
+#ifdef ECP_SM2P256_ASM
+/* Returns optimized methods for SM2 */
+const EC_METHOD *EC_GFp_sm2p256_method(void);
+#endif
+
 int ossl_ec_curve_nid_from_params(const EC_GROUP *group, BN_CTX *ctx);
 
 /* EC_METHOD definitions */