This patch optimizes SM2 for ARM processor using A64 instruction and precomputation table, which can speed up SM2 sign about 10 times and SM2 verify about 3 times. A new configure option `no-sm2-precomp` has been added to disable the precomputed table for point multiplicatin of the base point.

Perf data on Kunpeng-920 2.6GHz hardware looks like this:

|        | sign    | verify  | sign/s  | verify/s |
|--------|---------|---------|---------|----------|
| Before | 0.0004s | 0.0004s | 2438.2  | 2460.9   |
| After  | 0.0000s | 0.0001s | 27137.4 | 7434.0   |
| After (no-sm2-precomp)  | 0.0001s   | 0.0002s   | 7745.5   | 4411.2|

Signed-off-by: Xu Yizhou [xuyizhou1@huawei.com](mailto:xuyizhou1@huawei.com)
<!--
Thank you for your pull request. Please review these requirements:

Contributors guide: https://github.com/openssl/openssl/blob/master/CONTRIBUTING.md

Other than that, provide a description above this comment if there isn't one already

If this fixes a GitHub issue, make sure to have a line saying 'Fixes #XXXX' (without quotes) in the commit message.
-->

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->
- [ ] documentation is added or updated