@paulidale 2023-06-02
> I wonder if this optimization should be excluded if OPENSSL_SMALL_FOOTPRINT is defined. Given the large table size.

It would be a bonus but aarch64 isn't normally associated with _small_ devices.