@docularxu 2023-07-06
I'm from Linaro and specializing in Arm platform optimization. The latest version from @xu-yi-zhou effectively addresses the concerns I had previously raised. Also, I verified this version on Apple silicon M2, and the results align with the performance improvements mentioned earlier:
> which can speed up SM2 sign about 10 times and SM2 verify about 3 times.

With these, I would recommend merging this patch into the master branch.