@docularxu 2023-07-28
The feedbacks/comments so far have been fixed by yizhou in the latest patchset.
I verified again on the following setup. This patchset effectively enhances the performance as for aarch64 platforms, as claimed by the committer. I saw Sign/s 10 times gain, Verify/s 2.x times gain.
 - OS : Ubuntu 22.04 (via Parallel VM)
 - CPU: aarch64 Apple M2 Pro
 - C Compiler:     Version 11.3.0 (at /usr/bin/gcc)
 - C++ Compiler:   Version 11.3.0 (at /usr/bin/g++)

With that, I would give my approval on this patch. Hope it helps maintainers to merge the PR.