@t8m 2023-08-09
I am still worried about the size of the precomputed table - its size is 512kBytes, that's definitely not something to ignore in regards to the size of the rest of libcrypto. My feeling is we should at least make it possible to disable adding this if desired - i.e. either respect OPENSSL_SMALL_FOOTPRINT or even add a completely new define to disable it.