@t8m 2023-08-09
> The size of the precomputed table in file `ecp_nistz256_table.c` is 148kBytes, and it is included in libcrypto by default and no define to disable it (correct me if I'm wrong). I was wondering why this precomputed table is acceptable, is it because the size or something else?

Well.. nobody asked for the define when that one was included. It is also significantly smaller than the SM2 precomputed table. However it would be of course nice to have that one disableable as well.