@t8m 2023-08-09
**OTC question: Is the 512 kB read-only data table acceptable as is or do we require a possibility to disable it?**