@paulidale 2023-08-11
I think that `OPENSSL_SMALL_FOOTPRINT` should apply here.

That doesn't preclude having another mechanism as well, `OPENSSL_SMALL_FOOTPRINT` impacts a number of other algorithms and areas (reducing space and performance).  I expect a lot of people don't use SM2 and wouldn't want to pay the `OPENSSL_SMALL_FOOTPRINT` costs.