@t8m 2023-08-15
**OTC: We need compile-time config option for this that is on by default. The increased size of libcrypto also needs to be documented in CHANGES.md.**