@t8m 2023-08-15
@xu-yi-zhou Could you please add a new compile-time config option for this as OTC requested? It would not be a define but a full option like `sm2` itself is. I.e., `sm2-precomp` or something like that.