@xu-yi-zhou 2023-08-16
> We need compile-time config option for this that is on by default. The increased size of libcrypto also needs to be documented in CHANGES.md.

 OK, I will do that as soon as possible.