@xu-yi-zhou 2023-08-18
Hi all, I'm trying to add an option `enable-sm2-precomp` to enable the precomputed table. Do I need to add this option in `.github/workflows/*.yml`?