@t8m 2023-08-18
Yes, please. Add `no-sm2-precomp` to run-checker-daily.yml. Although it won't test much as this is arm specific and run-checker-daily runs on x86_64, it will at least test the disable-ability of the option.