@tom-cosgrove-arm 2023-08-21
> OTC: We need compile-time config option for this that is on by default. The increased size of libcrypto also needs to be documented in CHANGES.md.

My understanding of this request is that the large table would be included in default AArch64 builds, but can be disabled for smaller builds by configuring with `no-sm2-precomp`.

This seems to only use the large table if `enable-sm2-precomp` is given - @t8m is that what you wanted?