@xu-yi-zhou 2023-08-21
@tom-cosgrove-arm I think you are right, and I also forgot to add the  increased size of libcrypto in CHANGES.md, I will fix these.