@sususu98 2024-01-11
> This patch optimizes SM2 for ARM processor using A64 instruction and precomputation table, which can speed up SM2 sign about 10 times and SM2 verify about 3 times. A new configure option `no-sm2-precomp` has been added to disable the precomputed table for point multiplicatin of the base point.
> 
> Perf data on Kunpeng-920 2.6GHz hardware looks like this:
> 
> sign	verify	sign/s	verify/s
> Before	0.0004s	0.0004s	2438.2	2460.9
> After	0.0000s	0.0001s	27137.4	7434.0
> After (no-sm2-precomp)	0.0001s	0.0002s	7745.5	4411.2
> Signed-off-by: Xu Yizhou [xuyizhou1@huawei.com](mailto:xuyizhou1@huawei.com)
> 
> ##### Checklist
> * [ ]  documentation is added or updated

Hello, I tested using the default compilation options on Kunpeng-920 2.6GHz hardware and obtained a signature performance of 25057 times/s and a signature verification performance of 7042 times/s. May I ask if you added other options to the data obtained during compilation