@xu-yi-zhou 2024-01-11
> Hello, I tested using the default compilation options on Kunpeng-920 2.6GHz hardware and obtained a signature performance of 25057 times/s and a signature verification performance of 7042 times/s. May I ask if you added other options to the data obtained during compilation

no other options, try using the `taskset `command to bind the CPU while testing?